Updated Football.java
Struggled to get the logic of the subscript variables right when calculating the highest and lowest values in the scores[] array. Couldn't figure out how to use the conditional if statement to compare the actual array values, while using the body statement to store just the index location associated with either highest or lowest. Glad I finally figured out a working solution ~